add shared templates
