+++ lib/wl_web/templates/accounts/user/index.html.eex
@@ -6,7 +6,7 @@
           <span class="h3"> USERS </span>
         </div>
       </div>
-      <%= render "_users_search_form.html", %{conn: @conn} %>
+      <%= render WlWeb.SharedTemplatesView, "_search_form.html", %{conn: @conn, action: Routes.user_path(@conn, :index)} %>
       <%= if Enum.empty?(@users.entries) do %>
         <div class="row my-5">
           <div class="col d-flex justify-content-center">
@@ -50,33 +50,7 @@
             </div>
           </div>
         <% end %>
-        <div class="row my-5">
-          <div class="col d-flex justify-content-center">
-            <% page_number = @users.page_number %>
-            <% page_size = @users.page_size %>
-            <% total_pages = @users.total_pages %>
-            <% total_entries = @users.total_entries %>
-            <% start_index = (page_size * page_number + 1) - page_size %>
-            <% end_index = min(page_size * page_number, total_entries) %>
-            <nav class = "text-center" aria-label="Users Pagination">
-              <ul class="pagination pagination-lg">
-                <li class="<%= get_previous_page_class(page_number, total_pages)%>">
-                  <%= link(to: Routes.user_path(@conn, :index, Map.merge(@conn.params, %{page: page_number - 1, page_size: page_size})), class: "page-link") do %>
-                    <span aria-hidden="true">&laquo;</span>
-                    <span class="sr-only">Previous</span>
-                  <% end %>
-                </li>
-                <li class="<%= get_next_page_class(page_number, total_pages)%>">
-                  <%= link(to: Routes.user_path(@conn, :index, Map.merge(@conn.params, %{page: page_number + 1, page_size: page_size})), class: "page-link") do%>
-                    <span aria-hidden="true">&raquo;</span>
-                    <span class="sr-only">Next</span>
-                  <% end %>
-                </li>
-              </ul>
-              <span class="text-muted"> <%= "#{start_index} - #{end_index}  из #{total_entries}" %> </span>
-            </nav>
-          </div>
-        </div>
+        <%= render WlWeb.SharedTemplatesView, "_pagination.html", %{conn: @conn, entities: @users, path: :user_path} %>
       <% end %>
     </div>
   </div>
