+++ lib/wl_web/templates/shared_templates/_pagination.html.eex
@@ -1,0 +1,27 @@
+<div class="row my-5">
+  <div class="col d-flex justify-content-center">
+    <% page_number = @entities.page_number %>
+    <% page_size = @entities.page_size %>
+    <% total_pages = @entities.total_pages %>
+    <% total_entries = @entities.total_entries %>
+    <% start_index = (page_size * page_number + 1) - page_size %>
+    <% end_index = min(page_size * page_number, total_entries) %>
+    <nav class = "text-center" aria-label="Pagination">
+      <ul class="pagination pagination-lg">
+        <li class="<%= get_previous_page_class(page_number, total_pages)%>">
+          <%= link(to: apply(Routes, @path, [@conn, :index, Map.merge(@conn.params, %{page: page_number - 1, page_size: page_size})]), class: "page-link") do %>
+            <span aria-hidden="true">&laquo;</span>
+            <span class="sr-only">Previous</span>
+          <% end %>
+        </li>
+        <li class="<%= get_next_page_class(page_number, total_pages)%>">
+          <%= link(to: apply(Routes, @path, [@conn, :index, Map.merge(@conn.params, %{page: page_number + 1, page_size: page_size})]), class: "page-link") do%>
+            <span aria-hidden="true">&raquo;</span>
+            <span class="sr-only">Next</span>
+          <% end %>
+        </li>
+      </ul>
+      <span class="text-muted"> <%= "#{start_index} - #{end_index}  из #{total_entries}" %> </span>
+    </nav>
+  </div>
+</div>
