+++ lib/wl_web/templates/shared_templates/_search_form.html.eex
@@ -1,4 +1,4 @@
-<%= form_for @conn, Routes.user_path(@conn, :index), [method: :get], fn form -> %>
+<%= form_for @conn, @action, [method: :get], fn form -> %>
   <div class="container">
     <div class="form-group row justify-content-center">
       <%= text_input form, :search, class: "col-form-label col-4", placeholder: "Search" %>
