+++ lib/wl_web/views/shared_templates_view.ex
@@ -1,0 +1,3 @@
+defmodule WlWeb.SharedTemplatesView do
+  use WlWeb, :view
+end
